refactor: Calculator 내부 메서드 접근제한자 변경
